Add appContext with analytics
