+++ packages/common/src/index.ts
@@ -1,5 +1,6 @@
 export * from './api'
 export * from './audius-query'
+export * from './context'
 export * from './models'
 export * from './utils'
 export * from './services'
